requirements.txt
absl-py==1.4.0
alembic==1.11.2
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
bidict==0.22.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==4.4.2
dnspython==2.4.2
docopt==0.6.2
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.37
flask-ngrok==0.0.25
flask-restx==0.5.1
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
flatbuffers==23.5.26
fonttools==4.42.0
greenlet==2.0.2
gTTS==2.3.2
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
Jinja2==3.0.2
joblib==1.3.2
jsmin==3.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lesscpy==0.15.1
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mediapipe==0.10.3
moviepy==1.0.3
mysql-connector==2.2.9
mysqlclient==2.2.0
numpy==1.25.2
opencv-contrib-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pipreqs==0.4.13
ply==3.11
proglog==0.1.10
protobuf==3.20.3
pycparser==2.21
pygame==2.5.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2023.3
rcssmin==1.1.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
scikit-learn==1.2.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sounddevice==0.4.6
SQLAlchemy==1.4.29
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.0.3
WTForms==3.0.0
xxhash==3.3.0
yarg==0.1.9
